Each optic has been moved to its own file
